Add files via upload
Bandcamp Lossless Files (WAV) have the following naming conventions:
\Folder\*ARTIST*-*ALBUM*-*SONG NAME*.wav

This script, given an artist directory [at the highest], will rename a file from:
\Folder\*#* - *ARTIST* - *ALBUM* - *SONG NAME*.wav
to
\Folder\*#* *Song Name*

In foobar2000, if I try to automatically write the song into metadata (which is what appears in the song selector), it would take a very long time to comb through 25GB manually. Now, with this shorter convention, the song titles will be much more condensed and easier to find.